+++ libs/grid-angular/src/lib/grid-angular.spec.ts
@@ -1,7 +1,0 @@
-import { gridAngular } from './grid-angular.js';
-
-describe('gridAngular', () => {
-  it('should work', () => {
-    expect(gridAngular()).toEqual('grid-angular');
-  });
-});